refactor(ui-services): move services to dedicated folders
- toast-service
- form-validator-service
- sidebar-service
