+++ src/app/services/ui/sidebar-service.spec.ts
